Move useForm into its own file
